backend/pyproject.toml
[project]
name = "OtenkiLive"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.38.10",
    "dependency-injector>=4.46.0",
    "fastapi[all]>=0.115.12",
    "ffmpeg-python>=0.2.0",
    "loguru>=0.7.3",
    "pynamodb>=6.0.2",
    "requests>=2.32.3",
    "yt-dlp>=2025.4.30",
    "aws-lambda-typing>=2.20.0",
    "types-requests>=2.32.0.20250328",
    "pyyaml>=6.0.2",
    "types-pyyaml>=6.0.12.20250516",
    "boto3-stubs[boto3,s3,dynamodb,sqs]>=1.38.10",
    "boto3-stubs-custom",
    #    "pytest-env>=1.1.5",
    "torchvision>=0.22.0",
    "sahi>=0.11.23",
    "torch>=2.7.0",
    "packaging>=25.0",
    "ultralytics>=8.3.152",
    "onnx>=1.18.0",
    "onnxruntime>=1.22.0",
]

[dependency-groups]
dev = [
    "memory-profiler>=0.61.0",
    "memray>=1.17.2",
    "moto>=5.1.4",
    "mypy>=1.15.0",
    "psutil>=7.0.0",
    "pytest>=8.3.5",
    "pytest-memray>=1.7.0",
    "ruff>=0.11.8",
    "types-psutil>=7.0.0.20250601",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

#[tool.pytest_env]
#ENV="test"

[tool.pytest.ini_options]
pythonpath = "app"
testpaths = ["tests",]
filterwarnings = ["ignore::DeprecationWarning:botocore.*"]

[tool.mypy]
plugins = ['pydantic.mypy']

[[tool.mypy.overrides]]
module = ["yt_dlp", "ffmpeg", "torchvision.*", "sahi.*", "onnxruntime"]
ignore_missing_imports = true

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]

ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191",  # indentation contains tabs
    "B904",  # raise ... from ... になっていない場合のエラーを無視
]

[tool.ruff.lint.per-file-ignores]

[tool.ruff.lint.isort]
known-third-party = ["fastapi", "pydantic", "starlette"]

[tool.uv.sources]
boto3-stubs-custom = { path = "vendored/boto3_stubs_custom-1.38.10-py3-none-any.whl" }
#torch = [
#    { index = "pytorch-cpu" },
#]
#torchvision = [
#    { index = "pytorch-cpu" },
#]

#[[tool.uv.index]]
#name = "pytorch-cpu"
#url = "https://download.pytorch.org/whl/cpu"
#explicit = true
